Split legacy graph file effect tests
