+++ docs/COMPLETION_AUDIT.md
@@ -3698,6 +3698,11 @@ and do not assume Phase 4 is ready without evidence.
   `build_graph_command_accepts_identifier_fallback_declared_env_read_with_unselected_targets`,
   and
   `build_graph_command_rejects_env_read_without_fallback_before_unselected_targets`.
+- Legacy `zen build-graph build.zen` file-read host-effect tests now live in
+  focused `declared`, `rejections`, and `unselected_targets` modules under
+  `tests/integration/cli_build/legacy_graph_command_host_effects/file_reads/`,
+  preserving deterministic-effect coverage while avoiding one large mixed test
+  file.
 - Effect checking, typed allocator semantics, actors in std integration,
   JSON/YAML IR boundaries, and broader build graph execution remain gated by
   `docs/V1_SPEC.md`.
