+++ tests/integration/cli_build/legacy_graph_command_host_effects.rs
@@ -1,6 +1,6 @@
 use std::process::Command;
 
-#[path = "legacy_graph_command_host_effects/file_reads.rs"]
+#[path = "legacy_graph_command_host_effects/file_reads/mod.rs"]
 mod file_reads;
 
 #[test]
