+++ tests/integration/cli_build/legacy_graph_command_host_effects/file_reads/mod.rs
@@ -1,0 +1,6 @@
+#[path = "declared.rs"]
+mod declared;
+#[path = "rejections.rs"]
+mod rejections;
+#[path = "unselected_targets.rs"]
+mod unselected_targets;
